extract aspectk expressions to independent module
