+++ aspectk-compiler/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
 
 dependencies {
     implementation(project(":aspectk-plugin-common"))
+    implementation(project(":aspectk-expression"))
     implementation(libs.kotlin.compiler.embeddable)
     compileOnly(libs.auto.service)
     ksp(libs.auto.service.ksp)
