--- aspectk-compiler/src/main/kotlin/com/github/kitakkun/aspectk/compiler/pointcut/expression/AspectKExpression.kt
@@ -1,3 +1,0 @@
-package com.github.kitakkun.aspectk.compiler.pointcut.expression
-
-interface AspectKExpression
