+++ aspectk-compiler/src/test/kotlin/com/github/kitakkun/aspectk/compiler/lexer/PointcutExpressionLexerTest.kt
@@ -1,8 +1,8 @@
 package com.github.kitakkun.aspectk.compiler.lexer
 
-import com.github.kitakkun.aspectk.compiler.pointcut.lexer.AspectKPointcutExpressionLexer
-import com.github.kitakkun.aspectk.compiler.pointcut.lexer.AspectKToken
-import com.github.kitakkun.aspectk.compiler.pointcut.lexer.AspectKTokenType
+import com.github.kitakkun.aspectk.expression.lexer.AspectKPointcutExpressionLexer
+import com.github.kitakkun.aspectk.expression.lexer.AspectKToken
+import com.github.kitakkun.aspectk.expression.lexer.AspectKTokenType
 import org.junit.Test
 import kotlin.test.assertEquals
