+++ aspectk-expression/build.gradle.kts
@@ -1,0 +1,9 @@
+@file:Suppress("DSL_SCOPE_VIOLATION")
+plugins {
+    alias(libs.plugins.kotlinJvm)
+}
+
+dependencies {
+    implementation(libs.kotlin.compiler.embeddable)
+    testImplementation(kotlin("test"))
+}
