+++ aspectk-expression/src/main/kotlin/com/github/kitakkun/aspectk/expression/AspectKExpression.kt
@@ -1,0 +1,3 @@
+package com.github.kitakkun.aspectk.expression
+
+interface AspectKExpression
