+++ aspectk-expression/src/main/kotlin/com/github/kitakkun/aspectk/expression/ExecutionExpression.kt
@@ -1,3 +1,3 @@
-package com.github.kitakkun.aspectk.compiler.pointcut.expression
+package com.github.kitakkun.aspectk.expression
 
 data class ExecutionExpression(val functionMatchingExpression: FunctionMatchingExpression) : AspectKExpression
