+++ settings.gradle.kts
@@ -16,6 +16,7 @@ include(":aspectk-compiler")
 include(":aspectk-gradle-plugin")
 include(":aspectk-annotations")
 include(":aspectk-plugin-common")
+include(":aspectk-expression")
 
 rootProject.name = "AspectK"
